Computed properties are pretty handy for caching, should look further into this one

+++ app/Livewire/ArticleList.php
@@ -3,6 +3,7 @@
 namespace App\Livewire;
 
 use App\Models\Article;
+use Livewire\Attributes\Computed;
 use Livewire\Attributes\Title;
 use Livewire\WithPagination;
 
@@ -13,8 +14,25 @@ class ArticleList extends AdminComponent
 
     public bool $showOnlyPublished = false;
     
+    #[Computed(persist: false)]
+    public function articles() {
+        $query = Article::query();
+
+        if ($this->showOnlyPublished) {
+            $query->where('published', 1);
+        }
+        
+        return $query->latest()->paginate(10, ['id', 'title', 'content'], pageName: 'articles-page');
+    }
+    
     public function delete(Article $article) {
+        if ($this->articles->count() < 10) {
+            throw new \Exception("Nope");
+        }
+        
         $article->delete();
+        unset($this->articles);
+        cache()->forget('published-count');
     }
 
     public function showAll() {
@@ -27,18 +45,8 @@ class ArticleList extends AdminComponent
         $this->resetPage(pageName: 'articles-page');
     }
 
-    public function render()
-    {
-        $query = Article::query();
-
-        if ($this->showOnlyPublished) {
-            $query->where('published', 1);
-        }
-        
-        return view('livewire.article-list', [
-            'articles' => $query
-                ->latest()
-                ->paginate(10, ['id', 'title', 'content'], pageName: 'articles-page'),
-        ]);
-    }
+    // public function render()
+    // {
+    //     return view('livewire.article-list');
+    // }
 }
